Moved all asset related components to asset dir
