+++ client/src/components/asset/Asset.ts
